+++ fabric-biome-api-v1/src/main/java/net/fabricmc/fabric/impl/biome/VariantTransformer.java
@@ -25,7 +25,7 @@ import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.layer.util.LayerRandomnessSource;
 
-import net.fabricmc.fabric.api.biomes.v1.OverworldClimate;
+import net.fabricmc.fabric.api.biome.v1.OverworldClimate;
 
 /**
  * Deals with picking variants for you.
