+++ fabric-biome-api-v1/src/main/java/net/fabricmc/fabric/mixin/biome/MixinAddHillsLayer.java
@@ -31,7 +31,7 @@ import net.minecraft.world.biome.layer.BiomeLayers;
 import net.minecraft.world.biome.layer.util.LayerRandomnessSource;
 import net.minecraft.world.biome.layer.util.LayerSampler;
 
-import net.fabricmc.fabric.api.biomes.v1.OverworldBiomes;
+import net.fabricmc.fabric.api.biome.v1.OverworldBiomes;
 import net.fabricmc.fabric.impl.biome.InternalBiomeData;
 import net.fabricmc.fabric.impl.biome.InternalBiomeUtils;
 import net.fabricmc.fabric.impl.biome.WeightedBiomePicker;
