+++ fabric-biome-api-v1/src/main/resources/assets/fabric-biome-api-v1/icon.png
