+++ fabric-biome-api-v1/src/main/resources/fabric.mod.json
@@ -1,11 +1,11 @@
 {
   "schemaVersion": 1,
-  "id": "fabric-biomes-v1",
-  "name": "Fabric Biomes (v1)",
+  "id": "fabric-biome-api-v1",
+  "name": "Fabric Biome API (v1)",
   "version": "${version}",
   "environment": "*",
   "license": "Apache-2.0",
-  "icon": "assets/fabric-biomes-v1/icon.png",
+  "icon": "assets/fabric-biome-api-v1/icon.png",
   "contact": {
     "homepage": "https://fabricmc.net",
     "irc": "irc://irc.esper.net:6667/fabric",
@@ -21,6 +21,6 @@
   },
   "description": "Hooks for adding biomes to the default world generator.",
   "mixins": [
-    "fabric-biomes-v1.mixins.json"
+    "fabric-biome-api-v1.mixins.json"
   ]
 }
