+++ fabric-biome-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,12 +1,12 @@
 {
   "schemaVersion": 1,
-  "id": "fabric-biomes-v1-testmod",
-  "name": "Fabric Biomes (v1) Test Mod",
+  "id": "fabric-biome-api-v1-testmod",
+  "name": "Fabric Biome API (v1) Test Mod",
   "version": "1.0.0",
   "environment": "*",
   "license": "Apache-2.0",
   "depends": {
-    "fabric-biomes-v1": "*"
+    "fabric-biome-api-v1": "*"
   },
   "entrypoints": {
     "main": [
